Role assumes running ansible as owner, add drupal_core_owner_become.
Not sure this is the best solution but the current deploy tasks assume
this role is run as the core_owner and lay down the files and run
composer as the wrong user.

This gives an option to run the deploy tasks as the desired user as long
as you also set the drupal_core_owner to the correct user.